+++ TrafficGenerator.h
@@ -13,7 +13,7 @@ class TrafficGenerator {
 		~TrafficGenerator () {}
 
 		void gen_unicast_sd (unsigned int * src, unsigned int * dest);
-		void gen_temporal_parameters (double * time, unsigned int LorM); 
+		void gen_temporal_parameters (double * time, double LorM); 
 		void gen_first_request ();
 		void gen_request (double systemTime);
